New method of doing Steam Callbacks
This is something that I was thinking of trying out for a while,
it involves registering all the callbacks internally and then passing
along the callback through a single delegate function.

I didn't end up using it in Steamworks.NET but I'm sure someone will
find a use for it elsewhere.

It could be improved quite a lot by rewriting CCallback so that it
doesn't register a new function for each Callback as that ends up being
quite messy.